[Rename] #37 : Product -> KorProduct 변경
