+++ client/app/src/main/java/com/example/healthc/presentation/food/product/kor_product/KorProductFragment.kt
@@ -5,37 +5,36 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.healthc.R
-import com.example.healthc.databinding.FragmentSearchProductBinding
-import com.example.healthc.presentation.food.product.kor_product.SearchProductViewModel.SearchProductUiEvent
-import com.example.healthc.presentation.food.product.kor_product.adapter.SearchProductAdapter
+import com.example.healthc.databinding.FragmentKorProductBinding
+import com.example.healthc.presentation.food.product.kor_product.KorProductViewModel.SearchProductUiEvent
+import com.example.healthc.presentation.food.product.kor_product.adapter.KorProductAdapter
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @AndroidEntryPoint
-class SearchProductFragment : Fragment() {
+class KorProductFragment : Fragment() {
 
-    private var _binding: FragmentSearchProductBinding? = null
+    private var _binding: FragmentKorProductBinding? = null
     private val binding get() = checkNotNull(_binding)
 
-    private val viewModel : SearchProductViewModel by viewModels()
-    private val args : SearchProductFragmentArgs by navArgs()
+    private val viewModel : KorProductViewModel by viewModels()
 
-    private lateinit var searchProductAdapter : SearchProductAdapter
+    private lateinit var korProductAdapter : KorProductAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_search_product, container, false)
+        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_kor_product, container, false)
         binding.viewModel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
         return binding.root
@@ -44,7 +43,6 @@ class SearchProductFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         observeData()
-        initViewModelState()
         initAdapter()
         initButton()
     }
@@ -56,38 +54,33 @@ class SearchProductFragment : Fragment() {
                     is SearchProductUiEvent.Unit -> {}
 
                     is SearchProductUiEvent.Success -> {
-                        searchProductAdapter.submitList(it.foodIngredient.body.items)
+                        korProductAdapter.submitList(it.foodIngredient.body.items)
                     }
 
                     is SearchProductUiEvent.Failure -> {
-                        // TODO ERROR
+                        Toast.makeText(requireContext(), "상품 정보를 가져오는데 실패하였습니다.",
+                            Toast.LENGTH_SHORT).show()
                     }
                 }
             }
             .launchIn(viewLifecycleOwner.lifecycleScope)
     }
 
-    private fun initViewModelState(){
-        viewModel.setProduct(args.category)
-        viewModel.getFoodProduct()
-    }
-
     private fun initAdapter(){
-        searchProductAdapter = SearchProductAdapter()
+        korProductAdapter = KorProductAdapter()
         binding.searchProductRecyclerView.layoutManager = LinearLayoutManager(requireContext())
-        binding.searchProductRecyclerView.adapter = searchProductAdapter
+        binding.searchProductRecyclerView.adapter = korProductAdapter
     }
 
     private fun initButton(){
-        binding.backToProductButton.setOnClickListener{
-            navigateToProduct()
+        binding.backToCameraButton.setOnClickListener{
+            navigateToCamera()
         }
     }
 
-    private fun navigateToProduct(){
+    private fun navigateToCamera(){
         lifecycleScope.launchWhenStarted {
-            val direction = SearchProductFragmentDirections
-                .actionSearchProductFragmentToProductFragment()
+            val direction = KorProductFragmentDirections.actionKorProductFragmentToCameraFragment()
             findNavController().navigate(direction)
         }
     }
